requirements.txt
streamlit
pandas
openai
plotly-express
matplotlib
google-generativeai